Fix timestamp accuracy issues by reading whole pes packet
The read() method in dltsream is re-written to parse whole pes packets
and return one each call. This keeps the pts and dts in sync with the
data being given to the decoder, rather than running ahead in the
bitstream to flush the picture out of the decoder. This also saves an
memcpy() between the decode class and the format class.